+++ Changelog.md
@@ -2,7 +2,7 @@
 
 ## 1.4.0
 
-- Provide contextual menu `New Camel file` in Explorer. For `Create Camel Route with XXX DSL`, it creates them in the right-clicked folder.
+- Provide contextual menu `New Camel file` in Explorer. For `Camel Route using XXX DSL` and `Kamelet`, it creates them in the right-clicked folder.
 - Provide pre-configured launch and task configuration to Camel debug a Camel Quarkus application when using `Create a Camel Quarkus project` command
 
 ## 1.3.0